+++ packages/core/src/components/Logo/types.ts
@@ -1,0 +1,5 @@
+export type LogoProps = {
+    showName?: boolean;
+    logo?: React.ReactElement;
+    name?: React.ReactElement;
+};