+++ packages/theme/src/core/helperAndErrorTextTooltip/index.ts
@@ -1,0 +1,8 @@
+import colors from '../colors';
+import { HelperAndErrorTextTooltipTheme } from './types';
+
+const helperAndErrorTextTooltip: HelperAndErrorTextTooltipTheme = {
+    errorIconColor: colors.red[500]
+};
+
+export default helperAndErrorTextTooltip;