+++ packages/theme/src/layout/types.ts
@@ -1,2 +1,3 @@
+export * from './header/types';
 export * from './sideNav/types';
 export * from './sidePanel/types';