+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.10.1...@medly-components/theme@2.11.0) (2021-08-16)
+
+
+### Features
+
+* **layout:** new header layout component ([#377](https://github.com/medly/medly-components/issues/377)) ([d38bd12](https://github.com/medly/medly-components/commit/d38bd12b5f903e9ee29d61b812afc8c17dc617d3))
+
+
+
+
+
+## [2.10.1](https://github.com/medly/medly-components/compare/@medly-components/theme@2.10.0...@medly-components/theme@2.10.1) (2021-08-13)
+
+**Note:** Version bump only for package @medly-components/theme
+
+
+
+
+
 # [2.10.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.9.0...@medly-components/theme@2.10.0) (2021-08-11)